remove scamacs
The problem isn't that bundles third-party libraries, the problem is
that this is it's purpose.  In the authors words: "There is little
original code in Scamacs. It is simply the confluence of three major
pieces of Scala related elisp: Ensime, Scala-Mode, and ECB amalgamated
on Saturday morning."  In my own words: I should never have added this
to the mirror as it isn't a package, it's a starter-kit (which was
last updated four years ago.