remove 41 previously shelved non-packages
These modules are not actually packages and should therefore not have
been added to the Emacsmirror.  Some of these modules contain personal
tools that are hardly useful to anyone but their authors.  Others had
the potential to become packages, but were abandoned before getting
close and/or are so under-documented that it is highly unlikely that
anyone but the author every used them.