remove 50 shelved packages
All of these packages failed to define the license under which they
are released.  This is a problem because it is said that all Elisp
packages have to be considered to derive from Emacs and therefore have
to be released under a compatible license.

We could attempt to contact the authors of all of these packages, as
we have done for hundreds others, but I have decided that for these
packages it is not worth the effort.

These packages have other issues beside failing to specify the
license.  Some libraries do not provide a feature and/or come with no
or very little documentation.  That makes it unlikely that anyone but
the author ever managed to use these libraries.  Or the code is more
than a little dubious, making it doubtful that even the author used
it.  Or no contact information can be found.  Etc.